Move sticky notes methods to separate file
