--- designer/client/src/types/stickyNote.ts
@@ -1,12 +1,0 @@
-import type { ComponentDefinition } from "./scenarioGraph";
-
-export const StickyNoteType = "StickyNoteNode";
-
-export const StickyNoteDefinition: ComponentDefinition = {
-    parameters: [],
-    returnType: null,
-    icon: `/assets/components/${StickyNoteType}.svg`,
-    docsUrl: null,
-    outputParameters: null,
-    label: "",
-};
